+++ plasma_framework/test/helpers/constants.js
@@ -14,10 +14,13 @@ const OUTPUT_TYPE = {
 const EMPTY_BYTES = '0x';
 const EMPTY_BYTES_32 = '0x0000000000000000000000000000000000000000000000000000000000000000';
 
+const CHILD_BLOCK_INTERVAL = 1000;
+
 module.exports = {
     EMPTY_BYTES,
     EMPTY_BYTES_32,
     PROTOCOL,
     TX_TYPE,
     OUTPUT_TYPE,
+    CHILD_BLOCK_INTERVAL,
 };
