+++ plasma_framework/test/helpers/utils.js
@@ -1,3 +1,6 @@
+const { UtxoPos } = require('./positions.js');
+const { CHILD_BLOCK_INTERVAL } = require('./constants.js');
+
 async function spentOnGas(receipt) {
     const tx = await web3.eth.getTransaction(receipt.transactionHash);
     return web3.utils.toBN(tx.gasPrice).muln(receipt.gasUsed);
@@ -27,9 +30,23 @@ function computeNormalOutputId(txBytes, outputIndex) {
     );
 }
 
+function getOutputId(txBytes, utxoPos) {
+    const inputUtxoPos = new UtxoPos(utxoPos);
+    const outputId = isDeposit(inputUtxoPos.blockNum)
+        ? computeDepositOutputId(txBytes, inputUtxoPos.outputIndex, inputUtxoPos.utxoPos)
+        : computeNormalOutputId(txBytes, inputUtxoPos.outputIndex);
+    return outputId;
+}
+
+function isDeposit(blockNum) {
+    return blockNum % CHILD_BLOCK_INTERVAL !== 0;
+}
+
 module.exports = {
     spentOnGas,
     buildOutputGuard,
     computeDepositOutputId,
     computeNormalOutputId,
+    getOutputId,
+    isDeposit,
 };
